[docker] Apply setcap cap_sys_nice for /usr/bin/nice
Because we are using an specified non-root USER in the container the use of
'--cap-add SYS_NICE' is blocked by invoking 'nice' as a non-root
user. Furthermore the setting of security limits for the USER does not effect
'nice' as the container shell is not invoked through PAM. To allow the USER
access to 'nice' we setcap the executable.