[container] Fix CMakePresets.json source specification for Docker
Commit ID bd88ba0 resolves the path used to reference CMakePresets.json for LLVM
and Alive2 in Apptainer. In Docker the use of workdir and the specification of
the entrypoint in relation to this results in the script path being specified as
".". JSON based entrypoint specifications do not elaborate ENV or ARG. To
resolve this, build-and-run.bash now determines the absolute path of the script
using realpath.